divided business logic from UI
